Release 1.0

+++ app/src/main/java/di_griz/vkinfiles/DownloadThread.java
@@ -1,6 +1,5 @@
 package di_griz.vkinfiles;
 
-import android.content.Context;
 import android.util.Log;
 
 import org.apache.commons.io.FileUtils;
@@ -11,12 +10,10 @@ import java.net.URL;
 public class DownloadThread extends Thread {
     static int counter = 0;
     private String url;
-    private  Context context;
     private File file;
 
-    DownloadThread(Context context, File downloadedFile, String url) {
+    DownloadThread(File downloadedFile, String url) {
         counter++;
-        this.context = context;
         this.url = url;
         this.file = downloadedFile;
     }
@@ -24,11 +21,19 @@ public class DownloadThread extends Thread {
     @Override
     public void run() {
         try {
-            FileUtils.copyURLToFile(new URL(url), file);
-            Log.d("Download", "Complete " + file.getName());
+            if (file.exists())
+                Log.d("Download", "Файл " + file.getName() + " уже существует");
+            else {
+                FileUtils.copyURLToFile(new URL(url), file);
+                Log.d("Download", "Complete " + file.getName());
+            }
         } catch (Exception error) {
-            Log.e("Download", error.getMessage() + "; " + error.getClass().toString());
+            Log.e("Download", String.format("File: %s (%s; %s)", file.getName(),
+                    error.getMessage(), error.getClass().toString()));
+            if (file.delete())
+                Log.e("File", file.getName() + " was deleted");
         }
         counter--;
+        MainActivity.onDownloadingFinish();
     }
 }
